Rename the folders to "database/dataExportImportAPI" and "database/dataManagementAPI".
