+++ src/lib/UI/domainComponents/pageBlocks/AddDemoCoffeeBeans_PageBlock.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { onMount } from "svelte";
   import { anyCoffeeBeans } from "$lib/database/coffeeBeansAPI";
-  import { fillDbWithDemoData } from "$lib/database/manageData/manageData";
+  import { fillDbWithDemoData } from "$lib/database/dataManagementAPI/manageDataAPI";
   import { anyRecipes } from "$lib/database/recipesAPI";
   import NoItemsYetP from "$lib/UI/domainComponents/lists/NoItemsYetP.svelte";
   import { addToast } from "$lib/UI/genericComponents/toasts/toastProvider";
