+++ src/lib/UI/pageComponents/other/ImportDataBlock.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { importDataFromJson } from "$lib/database/manageImportExport/manageImportExport";
+  import { importDataFromJson } from "$lib/database/dataExportImportAPI/importDataAPI";
   import { isNullOrUndefined, isPresent } from "$lib/helpers/undefinedHelpers";
   import { addToast } from "$lib/UI/genericComponents/toasts/toastProvider";
