+++ src/lib/database/dataExportImportAPI/parse/recipes.ts
@@ -1,4 +1,4 @@
-import type { ImportedJsonRecipe } from "$lib/database/manageImportExport/types/ImportJSON";
+import type { ImportedJsonRecipe } from "$lib/database/dataExportImportAPI/types/ImportJSON";
 import { Recipe, type IRecipe } from "$lib/domain/entities/Recipe";
 import { isNullOrUndefined } from "$lib/helpers/undefinedHelpers";
 import {
